Add: 장바구니 비어 있는 상태와 장바구니 아이템 담겨 있을 때 분리
